+++ src/poly/vector.h
@@ -7,10 +7,10 @@ namespace hc {
 
 class Vector {
 private:
-	double x_, y_;
+	float x_, y_;
 
 public:
-	Vector(double x, double y);
+	Vector(float x, float y);
 	~Vector();
 
 	// Algebric operations
@@ -24,30 +24,30 @@ public:
 
 
 	Vector operator*(int k);
-	Vector operator*(double k);
+	Vector operator*(float k);
 
-	double operator*(Vector vec);
-	double operator%(Vector vec);
+	float operator*(Vector vec);
+	float operator%(Vector vec);
 
 	// Projection of this vector on vector 'vec'
 	Vector proj(Vector vec);
 
 	// Setting components
-	void set_x(double x) { x_ = x; }
-	void set_y(double y) { y_ = y; }
+	void set_x(float x) { x_ = x; }
+	void set_y(float y) { y_ = y; }
 
 	// Setting through geometric values. These have a longer
 	// implementation, and can be found in the .cc file
-	void set_norm(double norm);
-	void set_angle(double angle);
+	void set_norm(float norm);
+	void set_angle(float angle);
 
 	// Reading componenents
-	double x() { return x_; }
-	double y() { return y_; }
+	float x() { return x_; }
+	float y() { return y_; }
 
 	// Reading geometric values
-	double norm() { return sqrt(pow(x_, 2) + pow(y_, 2)); }
-	double angle() { return atan2(y_, x_); }
+	float norm() { return sqrt(pow(x_, 2) + pow(y_, 2)); }
+	float angle() { return atan2(y_, x_); }
 
 #ifdef _DEBUG
 	void print();
